:clown_face: fake data

+++ service-booking/src/main/java/com/tamnguyen/servicebooking/models/Flight.java
@@ -47,7 +47,7 @@ public class Flight {
   private LocalDateTime arrivalTime;
 
   @NotNull
-  private BigDecimal ticketPrice;
+  private Double ticketPrice;
 
   @NotNull
   private String aircraftType;
